Home screen
making changes according to comments
