--- app/src/main/java/care/intouch/app/feature/home/di/HomeDI.kt
@@ -1,12 +1,0 @@
-package care.intouch.app.feature.home.di
-
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-object HomeDI {
-
-}
-
