+++ app/src/main/java/care/intouch/app/feature/home/presentation/models/HomeScreenSideEffect.kt
@@ -4,14 +4,11 @@ import care.intouch.uikit.common.StringVO
 
 sealed interface HomeScreenSideEffect {
     data class ShowDialog(
-        val title: StringVO,
-        val onConfirm: () -> Unit,
-        val onDismiss: () -> Unit,
+        val title: StringVO = StringVO.Plain(""),
+        val massage: StringVO = StringVO.Plain(""),
+        val onConfirmButtonText: StringVO = StringVO.Plain(""),
+        val onDismissButtonText: StringVO = StringVO.Plain(""),
+        val onConfirm: () -> Unit = {},
+        val onDismiss: () -> Unit = {},
     ) : HomeScreenSideEffect
-
 }
-data class DialogState(
-    val title: StringVO = StringVO.Plain(""),
-    val onConfirm: () -> Unit = {},
-    val onDismiss: () -> Unit = {}
-)
